Crate r2r_actions

Re-exports

pub use self::rcl_action_goal_event_e as rcl_action_goal_event_t;

Structs

__BindgenOpaqueArray

If Bindgen could only determine the size and alignment of a type, it is represented like this.

rcl_action_cancel_response_s

Enums

rcl_action_goal_event_e

Constants

RCL_RET_ACTION_CLIENT_INVALID

RCL_RET_ACTION_CLIENT_TAKE_FAILED

RCL_RET_ACTION_GOAL_ACCEPTED

RCL_RET_ACTION_GOAL_EVENT_INVALID

RCL_RET_ACTION_GOAL_HANDLE_INVALID

RCL_RET_ACTION_GOAL_REJECTED

RCL_RET_ACTION_NAME_INVALID

RCL_RET_ACTION_SERVER_INVALID

RCL_RET_ACTION_SERVER_TAKE_FAILED

Functions

rcl_action_accept_new_goal⚠

rcl_action_cancel_response_fini⚠

rcl_action_cancel_response_init⚠

rcl_action_client_fini⚠

rcl_action_client_get_action_name⚠

rcl_action_client_get_default_options⚠

rcl_action_client_get_options⚠

rcl_action_client_init⚠

rcl_action_client_is_valid⚠

rcl_action_client_set_cancel_client_callback⚠

rcl_action_client_set_feedback_subscription_callback⚠

rcl_action_client_set_goal_client_callback⚠

rcl_action_client_set_result_client_callback⚠

rcl_action_client_set_status_subscription_callback⚠

rcl_action_client_wait_set_get_entities_ready⚠

rcl_action_client_wait_set_get_num_entities⚠

rcl_action_expire_goals⚠

rcl_action_get_client_names_and_types_by_node⚠

rcl_action_get_goal_status_array⚠

rcl_action_get_names_and_types⚠

rcl_action_get_server_names_and_types_by_node⚠

rcl_action_get_zero_initialized_cancel_request⚠

rcl_action_get_zero_initialized_cancel_response⚠

rcl_action_get_zero_initialized_client⚠

rcl_action_get_zero_initialized_goal_handle⚠

rcl_action_get_zero_initialized_goal_info⚠

rcl_action_get_zero_initialized_goal_status_array⚠

rcl_action_get_zero_initialized_server⚠

rcl_action_goal_handle_fini⚠

rcl_action_goal_handle_get_info⚠

rcl_action_goal_handle_get_status⚠

rcl_action_goal_handle_init⚠

rcl_action_goal_handle_is_active⚠

rcl_action_goal_handle_is_cancelable⚠

rcl_action_goal_handle_is_valid⚠

rcl_action_goal_status_array_fini⚠

rcl_action_goal_status_array_init⚠

rcl_action_notify_goal_done⚠

rcl_action_process_cancel_request⚠

rcl_action_publish_feedback⚠

rcl_action_publish_status⚠

rcl_action_send_cancel_request⚠

rcl_action_send_cancel_response⚠

rcl_action_send_goal_request⚠

rcl_action_send_goal_response⚠

rcl_action_send_result_request⚠

rcl_action_send_result_response⚠

rcl_action_server_fini⚠

rcl_action_server_get_action_name⚠

rcl_action_server_get_default_options⚠

rcl_action_server_get_goal_handles⚠

rcl_action_server_get_options⚠

rcl_action_server_goal_exists⚠

rcl_action_server_init⚠

rcl_action_server_is_available⚠

rcl_action_server_is_valid⚠

rcl_action_server_is_valid_except_context⚠

rcl_action_server_set_cancel_service_callback⚠

rcl_action_server_set_goal_service_callback⚠

rcl_action_server_set_result_service_callback⚠

rcl_action_server_wait_set_get_entities_ready⚠

rcl_action_server_wait_set_get_num_entities⚠

rcl_action_take_cancel_request⚠

rcl_action_take_cancel_response⚠

rcl_action_take_feedback⚠

rcl_action_take_goal_request⚠

rcl_action_take_goal_response⚠

rcl_action_take_result_request⚠

rcl_action_take_result_response⚠

rcl_action_take_status⚠

rcl_action_transition_goal_state⚠

rcl_action_update_goal_state⚠

rcl_action_wait_set_add_action_client⚠

rcl_action_wait_set_add_action_server⚠

Type Aliases

rcl_action_cancel_request_t

rcl_action_cancel_response_t

rcl_action_client_options_t

rcl_action_client_t

rcl_action_goal_handle_t

rcl_action_goal_info_t

rcl_action_goal_state_t

rcl_action_goal_status_array_t

rcl_action_server_options_t

rcl_action_server_t